Enum svg::node::element::path::Command

pub enum Command {
    Move(Position, Parameters),
    Line(Position, Parameters),
    HorizontalLine(Position, Parameters),
    VerticalLine(Position, Parameters),
    QuadraticCurve(Position, Parameters),
    SmoothQuadraticCurve(Position, Parameters),
    CubicCurve(Position, Parameters),
    SmoothCubicCurve(Position, Parameters),
    EllipticalArc(Position, Parameters),
    Close,
}

A command of a data attribute.

Variants
Move(Position, Parameters)

Establish a new current point.

Line(Position, Parameters)

Draw straight lines.

HorizontalLine(Position, Parameters)

Draw horizontal lines.

VerticalLine(Position, Parameters)

Draw vertical lines.

QuadraticCurve(Position, Parameters)

Draw a quadratic Bézier curve.

SmoothQuadraticCurve(Position, Parameters)

Draw a quadratic Bézier curve assuming the control point to be the reflection of the control point on the previous command relative to the current point.

CubicCurve(Position, Parameters)

Draw a cubic Bézier curve.

SmoothCubicCurve(Position, Parameters)

Draw a cubic Bézier curve assuming the first control point to be the reflection of the second control point on the previous command relative to the current point.

EllipticalArc(Position, Parameters)

Draw an elliptical arc.

Close

End the current subpath.